refactor: split retriever.ts into focused search modules
Extract the 1895-line retriever.ts into dedicated search modules under
assistant/src/memory/search/:

- types.ts: Shared types (Candidate, CandidateSource, MemoryRecallResult, etc.)
- lexical.ts: FTS-based search, recency search, direct item LIKE search
- semantic.ts: Qdrant vector search, cosine mapping, connection error detection
- entity.ts: Entity matching, BFS relation traversal (with cycle detection),
  entity-linked item retrieval
- ranking.ts: RRF merge, source caps, freshness decay, trust weighting,
  LLM re-ranking, token budget trimming, item usage tracking
- formatting.ts: Injection text formatting (markdown/structured_v1),
  attention ordering, XML escaping, time formatting

The main retriever.ts is now the orchestrator (~530 lines) that imports
from search modules and re-exports the public API, keeping all existing
importers working without changes.
